+++ src/main/java/com/kh/mng/location/model/vo/DetailLocation_.java
@@ -2,9 +2,6 @@ package com.kh.mng.location.model.vo;
 
 import java.sql.Date;
 import java.sql.Time;
-import java.util.ArrayList;
-
-import com.kh.mng.common.model.vo.Attachment;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -17,7 +14,7 @@ import lombok.ToString;
 @Getter
 @Setter
 @ToString
-public class DetailLocation2 {
+public class DetailLocation {
 	private int locationNo;//장소 번호
 	private String locationName;//장소 이름
 	private String locationPhone;//장소 전화번호
@@ -25,26 +22,16 @@ public class DetailLocation2 {
 	private double locationStar;//장소 평점
 	private String reservationLink;//예약 링크
 	private String address;//장소 주소
+	private String goods;//상품이름
+	private String goodPrice;//상품가격
+	private String roomInfo;//객실 정보
+	private String day; //체크인 체크아웃
+    private int capacity;//수용인원
 	private int locationCategoryNo;//카테고리 번호
 	private String categoryName;//카테고리 명
-	private ArrayList<LocationOption> locationOption; //장소 옵션들
-    private ArrayList<Attachment> attachMent; //첨부파일들
-    private ArrayList<OperationTime>operationTime;//운영시간
-    
-    //출입등급,반려동물크기,반려동물 분류를 조인시킨 값들
-    private ArrayList<PetKindGrade> petKindNo;
-  
-
-   
-    
-    
-    
-    
-    
-
-
-	
-	
-	
+	private Time startTime;//시작시간
+	private Time endTime;//종료시간
+	private boolean restStatus;//휴무여부
+	private String kindName;//분류 명
 	
 }
