+++ src/main/resources/mybatis-config.xml
@@ -15,13 +15,17 @@
  		<typeAlias type="com.kh.mng.member.model.vo.Member" alias="Member" />
  		<typeAlias type="com.kh.mng.pet.model.vo.Pet" alias="Pet" />
  		<typeAlias type="com.kh.mng.location.model.vo.Picked" alias="Picked"/>
- 		<typeAlias type="com.kh.mng.location.model.vo.DetailLocation" alias="DetailLocation"/>
+ 		<typeAlias type="com.kh.mng.location.model.vo.DetailLocation_" alias="DetailLocation_"/>
  		<typeAlias type="com.kh.mng.location.model.vo.Review" alias="Review"/>
  		<typeAlias type="com.kh.mng.location.model.vo.DetailLocationAttachment" alias="DetailLocationAttachment"/>
  		<typeAlias type="com.kh.mng.common.model.vo.Attachment" alias="Attachment"/>
  		<typeAlias type="com.kh.mng.location.model.vo.EnterGrade" alias="EnterGrade"/>
  		<typeAlias type="com.kh.mng.location.model.vo.Location" alias="Location"/>
  		<typeAlias type="com.kh.mng.community.model.vo.Board" alias="Board"/>
+ 		<typeAlias type="com.kh.mng.location.model.vo.DetailLocation" alias="DetailLocation"/>
+ 		<typeAlias type="com.kh.mng.location.model.vo.LocationOption" alias="LocationOption"/>
+ 		<typeAlias type="com.kh.mng.location.model.vo.OperationTime" alias="OperationTime"/>
+ 		<typeAlias type="com.kh.mng.location.model.vo.PetKindGrade" alias="PetkindGrade"/>
  		<!-- <typeAlias type="com.kh.spring.board.model.vo.Board" alias="Board" />
  		<typeAlias type="com.kh.spring.common.model.vo.Reply" alias="Reply" /> -->
  		
@@ -39,6 +43,7 @@
  		<mapper resource="/mappers/location/attachment-mapper.xml"/>
  		<mapper resource="/mappers/search-mapper.xml"/>
  		<mapper resource="/mappers/main-mapper.xml"/>
+ 		<mapper resource="/mappers/location/location-detail-mapper.xml"/>
  	</mappers>
